refactor register service
