--- src/api/api.js
@@ -1,1 +1,0 @@
-export const api = "http://n2j-api.herokuapp.com";
